KMRL-Front-AI/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
faiss-cpu>=1.9.0
transformers>=4.35.2
sentence-transformers>=2.2.2
pytesseract>=0.3.10
opencv-python>=4.8.1
celery>=5.3.4
redis>=5.0.1
neo4j>=5.14.1
pandas>=2.1.3
numpy>=1.25.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
Pillow>=10.0.0